Add files via upload
Bokeh 
Bokeh is a data visualization library in Python that provides high-performance interactive charts and plots. Bokeh output can be obtained in various mediums like notebook, html and server. It is possible to embed bokeh plots in Django and flask apps. 
 
Bokeh provides two visualization interfaces to users: 
bokeh.models : A low level interface that provides high flexibility to application developers. bokeh.plotting : A high level interface for creating visual glyphs. 
The major concept of Bokeh is that graphs are built up one layer at a time. We start out by creating a figure, and then we add elements, called glyphs, to the figure. (For those who have used ggplot, the idea of glyphs is essentially the same as that of geoms which are added to a graph one ‘layer’ at a time.)  
Glyphs can take on many shapes depending on the desired use: circles, lines, patches, bars, arcs, and so on. Let’s illustrate the idea of glyphs by making a basic chart with squares and circles. First, we make a plot using the figure method and then we append our glyphs to the plot by calling the appropriate method and passing in data.  
Bokeh is a web-focused tool for creating interactive plots. It supports streaming datasets and integrates with Pandas by using the ColumnDataSource class. There are built-in tools that can be included on a widget box attached to the plot and used to explore the data in an interactive way, such as zooming in, selecting and overlaying a crosshair. 
Plotting with this tool is different because it is built around glyphs. bokeh.models is a lowlevel interface that implements glyphs like Line and Circle, which are visual shapes with properties attached to the data, like its coordinates, size, and color. The higher-level interface bokeh.plotting offers methods to display glyphs, but it is not as abstracted as one may expect. To plot a histogram, for example, the edges of each bin must be calculated to compose two lists, which are passed to plotting method quad that displays Quad glyphs. The obtained plot is pretty and interactive, but a lot of manual work is required. 
Bokeh server is a Flask Blueprint for building interactive web applications. This can also be accomplished with bokeh.embed, whose methods can translate plots into JavaScript and HTML components. Both allow callback functions, bringing interaction to another level by reacting to sliders or any other information input supported by HTML. 
When plotting multiple plots with the same range, Bokeh synchronizes them so interactions made in one can affect the other ones. This is a really useful tool for exploring and comparing details of the data. 
Pandas 
Pandas is a software library written for the Python programming language. It is used for data manipulation and analysis. It provides special data structures and operations for the manipulation of numerical tables and time series. Pandas is an open-source library that is built on top of NumPy library. It is mainly popular for importing and analyzing data much easier. Pandas is fast and it has high-performance & productivity for users.